+++ src/services/leaderboard-stat-getter.js
@@ -1,6 +1,7 @@
 import _ from 'lodash';
 import { Schemas } from 'forest-express';
 import BaseStatGetter from './base-stat-getter';
+import Orm from '../utils/orm';
 
 function LeaderboardStatGetter(model, modelRelationship, params, options) {
   BaseStatGetter.call(this, model, params, options);
@@ -25,7 +26,7 @@ function LeaderboardStatGetter(model, modelRelationship, params, options) {
     // NOTICE: As MySQL cannot support COUNT(table_name.*) syntax, fieldName cannot be '*'.
     const fieldName = aggregateField || schemaRelationship.primaryKeys[0] ||
       schemaRelationship.fields[0].field;
-    return `${schemaRelationship.name}.${fieldName}`;
+    return `${schemaRelationship.name}.${Orm.getColumnName(schema, fieldName)}`;
   }
 
   this.perform = () => modelRelationship