+++ src/services/line-stat-getter.js
@@ -4,6 +4,7 @@ import moment from 'moment';
 import { Schemas } from 'forest-express';
 import BaseStatGetter from './base-stat-getter';
 import { isMySQL, isMSSQL, isSQLite } from '../utils/database';
+import Orm from '../utils/orm';
 
 // jshint sub: true
 function LineStatGetter(model, params, options) {
@@ -15,14 +16,14 @@ function LineStatGetter(model, params, options) {
   function getAggregateField() {
     // NOTICE: As MySQL cannot support COUNT(table_name.*) syntax, fieldName
     //         cannot be '*'.
-    const fieldName = params.aggregate_field || schema.primaryKeys[0] ||
-      schema.fields[0].field;
-    return `${schema.name}.${fieldName}`;
+    const fieldName = params.aggregate_field
+      || schema.primaryKeys[0]
+      || schema.fields[0].field;
+    return `${schema.name}.${Orm.getColumnName(schema, fieldName)}`;
   }
 
   function getGroupByDateField() {
-    const fieldName = params.group_by_date_field;
-    return `${schema.name}.${fieldName}`;
+    return `${schema.name}.${Orm.getColumnName(schema, params.group_by_date_field)}`;
   }
 
   const groupByDateField = getGroupByDateField();